Add missing errors
These are tricky. A `<math>` tag as the first token in a document
fragment whose context node is `td` (or `th`, but this isn't tested
here) is fine. One in a document fragment whose context node is `tr`,
`thead`, `tbody`, or `tfoot` is a parse error and the elements are
foster parented.

The table element start tags after the `<mo>` tag are parsed as html and
there are no `tr` (for `tr` elements) elements or `thead`, `tbody`, or
`tfoot` (for the others) elements in table scope which is a parse error.
I just invented some sames for those.

The `</table>` tag after the `<mo>` is parsed as foreign but it doesn't
match `<mo>` so it's a parse error.

Finally, the EOF occurs while a bunch of elements are open which is a
parse error.